@@ -13,21 +13,19 @@ class ViewRecitation extends StatefulWidget {
   final SnackbarNeeded snackbarNeeded;
 
   const ViewRecitation(
-      {Key key, this.narration, this.loadingStateChanged, this.snackbarNeeded})
+      {Key? key,
+      required this.narration,
+      required this.loadingStateChanged,
+      required this.snackbarNeeded})
       : super(key: key);
 
   @override
-  State<StatefulWidget> createState() => _ViewRecitationState(
-      );
+  State<StatefulWidget> createState() => _ViewRecitationState();
 }
 
 class _ViewRecitationState extends State<ViewRecitation>
     with AfterLayoutMixin<ViewRecitation> {
- 
-
-  AudioPlayer _player;
-
-
+  AudioPlayer? _player;
 
   final _titleController = TextEditingController();
   final _artistNameController = TextEditingController();
@@ -44,7 +42,7 @@ class _ViewRecitationState extends State<ViewRecitation>
 
   @override
   void dispose() {
-    _player.dispose();
+    _player!.dispose();
     _titleController.dispose();
     _artistNameController.dispose();
 
@@ -52,14 +50,11 @@ class _ViewRecitationState extends State<ViewRecitation>
   }
 
   String getVerse(PublicRecitationViewModel narration, Duration position) {
-    if (position == null || narration == null || narration.verses == null) {
+    if (narration.verses == null) {
       return '';
     }
-    var verse = narration.verses.lastWhere(
+    var verse = narration.verses!.lastWhere(
         (element) => element.audioStartMilliseconds <= position.inMilliseconds);
-    if (verse == null) {
-      return '';
-    }
     return verse.verseText;
   }
 
@@ -68,11 +63,10 @@ class _ViewRecitationState extends State<ViewRecitation>
 
   @override
   Widget build(BuildContext context) {
-    _titleController.text = widget.narration == null ? '' : widget.narration.poemFullTitle;
-    _artistNameController.text = widget.narration == null ? '' : widget.narration.audioArtist;
-    _fileDownloadTitle = widget.narration == null
-        ? ''
-        : 'دریافت با حجم ${(widget.narration.mp3SizeInBytes / (1024 * 1024)).toStringAsFixed(2)} مگابایت';
+    _titleController.text = widget.narration.poemFullTitle;
+    _artistNameController.text = widget.narration.audioArtist;
+    _fileDownloadTitle =
+        'دریافت با حجم ${(widget.narration.mp3SizeInBytes / (1024 * 1024)).toStringAsFixed(2)} مگابایت';
 
     return FocusTraversalGroup(
         child: Form(
@@ -124,14 +118,14 @@ class _ViewRecitationState extends State<ViewRecitation>
                   crossAxisAlignment: CrossAxisAlignment.center,
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
-                    ControlButtons(_player, widget.narration, widget.loadingStateChanged,
-                        widget.snackbarNeeded),
-                    StreamBuilder<Duration>(
-                      stream: _player.durationStream,
+                    ControlButtons(_player!, widget.narration,
+                        widget.loadingStateChanged, widget.snackbarNeeded),
+                    StreamBuilder<Duration?>(
+                      stream: _player!.durationStream,
                       builder: (context, snapshot) {
                         final duration = snapshot.data ?? Duration.zero;
                         return StreamBuilder<Duration>(
-                          stream: _player.positionStream,
+                          stream: _player!.positionStream,
                           builder: (context, snapshot) {
                             var position = snapshot.data ?? Duration.zero;
                             if (position > duration) {
@@ -145,7 +139,7 @@ class _ViewRecitationState extends State<ViewRecitation>
                                     duration: duration,
                                     position: position,
                                     onChangeEnd: (newPosition) {
-                                      _player.seek(newPosition);
+                                      _player!.seek(newPosition);
                                     },
                                   ),
                                   Text(getVerse(widget.narration, position))
