@@ -15,7 +15,7 @@ class ControlButtons extends StatelessWidget {
 
   const ControlButtons(this.player, this.narration, this.loadingStateChanged,
       this.snackbarNeeded,
-      {Key key})
+      {Key? key})
       : super(key: key);
 
   @override
@@ -113,13 +113,13 @@ class ControlButtons extends StatelessWidget {
 class SeekBar extends StatefulWidget {
   final Duration duration;
   final Duration position;
-  final ValueChanged<Duration> onChanged;
-  final ValueChanged<Duration> onChangeEnd;
+  final ValueChanged<Duration>? onChanged;
+  final ValueChanged<Duration>? onChangeEnd;
 
   const SeekBar({
-    Key key,
-    @required this.duration,
-    @required this.position,
+    Key? key,
+    required this.duration,
+    required this.position,
     this.onChanged,
     this.onChangeEnd,
   }) : super(key: key);
@@ -129,7 +129,7 @@ class SeekBar extends StatefulWidget {
 }
 
 class _SeekBarState extends State<SeekBar> {
-  double _dragValue;
+  double? _dragValue;
 
   @override
   Widget build(BuildContext context) {
@@ -145,12 +145,12 @@ class _SeekBarState extends State<SeekBar> {
               _dragValue = value;
             });
             if (widget.onChanged != null) {
-              widget.onChanged(Duration(milliseconds: value.round()));
+              widget.onChanged!(Duration(milliseconds: value.round()));
             }
           },
           onChangeEnd: (value) {
             if (widget.onChangeEnd != null) {
-              widget.onChangeEnd(Duration(milliseconds: value.round()));
+              widget.onChangeEnd!(Duration(milliseconds: value.round()));
             }
             _dragValue = null;
           },
@@ -163,7 +163,7 @@ class _SeekBarState extends State<SeekBar> {
                       .firstMatch('$_remaining')
                       ?.group(1) ??
                   '$_remaining',
-              style: Theme.of(context).textTheme.caption),
+              style: Theme.of(context).textTheme.bodySmall),
         ),
       ],
     );
@@ -173,17 +173,17 @@ class _SeekBarState extends State<SeekBar> {
 }
 
 _showSliderDialog({
-  BuildContext context,
-  String title,
-  int divisions,
-  double min,
-  double max,
+  BuildContext? context,
+  required String title,
+  required int divisions,
+  required double min,
+  required double max,
   String valueSuffix = '',
-  Stream<double> stream,
-  ValueChanged<double> onChanged,
+  Stream<double>? stream,
+  ValueChanged<double>? onChanged,
 }) {
   showDialog(
-    context: context,
+    context: context!,
     builder: (context) => AlertDialog(
       title: Text(title, textAlign: TextAlign.center),
       content: StreamBuilder<double>(
